refactor: ADR-004 Phase 2 - Module-Core Boundary Establishment
## Changes

### Event Migration (Unit 1-4)
- Move ExpectationCalculationCompletedEvent to module-core
- Update all imports in production and test code

### Unused Class Cleanup (Unit 5-6)
- Delete unused Java classes: FetchResult, CompensationCommand
- Delete deprecated PMF and policy classes: DensePmf, SparsePmf, CostCalculationStrategy, TableBasedCostStrategy
- Delete corresponding test files

### Interface Migration (Unit 7)
- Delete deprecated PersistenceTrackerStrategy.java (was extending core interface)
- Update EquipmentPersistenceTracker and RedisEquipmentPersistenceTrackerAdapter to use core interface directly

## Verification
- All compilation errors resolved
- All tests pass
- module-core only depends on module-common
