+++ docs/05_Reports/session-report-2026-03-02-214458.md
@@ -1,0 +1,49 @@
+# 세션 리포트
+
+**생성 일시**: 2026-03-02 21:44:59
+**브랜치**: feature/outbox-shutdown-infra
+
+## 최근 커밋
+
+```
+a7a978d fix: OutboxProcessorTest shouldCallHandleFailureOnProcessingError 테스트 수정
+33387e2 docs: ADR-010 Service Layer 모듈화 전략 수립 (#531)
+7899315 refactor: module-infra, module-web Java → Kotlin 변환 (#529)
+dd7d7b3 fix: ADR-005 CursorPageResponse를 common에서 page 패키지로 이동 (#527)
+0757865 fix: ADR-005 CI 컴파일 에러 수정 및 테스트 import 업데이트 (#525)
+```
+
+## 변경된 파일 (최근 5개 커밋)
+
+```
+Note
+docs/05_Reports/session-report-2026-03-01-111627.md
+docs/05_Reports/session-report-2026-03-01-141058.md
+docs/05_Reports/session-report-2026-03-01-141954.md
+docs/05_Reports/session-report-2026-03-01-153000.md
+docs/05_Reports/session-report-2026-03-01-171619.md
+docs/05_Reports/session-report-2026-03-01-194833.md
+docs/05_Reports/session-report-2026-03-01-202118.md
+docs/05_Reports/session-report-2026-03-01-213359.md
+docs/05_Reports/session-report-2026-03-01-220949.md
+docs/05_Reports/session-report-2026-03-02-135835.md
+docs/05_Reports/session-report-2026-03-02-144854.md
+docs/05_Reports/session-report-2026-03-02-154638.md
+docs/05_Reports/session-report-2026-03-02-154857.md
+docs/adr/009-cache-to-infra-migration.md
+docs/adr/010-service-layer-modularization.md
+docs/adr/011-donation-outbox-infra-migration.md
+docs/adr/ADR-005-monitoring-infra-migration.md
+module-app/src/main/java/maple/expectation/adapter/in/DlqPortAdapter.java
+module-app/src/main/java/maple/expectation/controller/dto/dlq/DlqDetailResponse.java
+```
+
+## 검증 결과
+
+- ✅ PfastTest 통과
+- ✅ DoD 검증 통과
+- ✅ Stateless 설계 준수
+- ✅ 미완성 코드 없음
+
+---
+*자동 생성됨 by stop-validation.sh*
