+++ docs/05_Reports/session-report-2026-03-03-000127.md
@@ -1,0 +1,49 @@
+# 세션 리포트
+
+**생성 일시**: 2026-03-03 00:01:27
+**브랜치**: develop
+
+## 최근 커밋
+
+```
+730dcef refactor: ADR-011 Cache/Worker module-infra 이관 (#534)
+93b0adb refactor: TotalExpectationCacheService를 module-infra로 이관 (#533)
+36e6c1e refactor: ADR-011 EquipmentFingerprintGenerator를 module-infra로 이관 (#532)
+c28b8dd refactor: ADR-008 Discord Alert Service를 module-infra로 이관 (#530)
+c0b37b5 refactor: ADR-012 like 패키지 core/infra 분리 완료 (#535)
+```
+
+## 변경된 파일 (최근 5개 커밋)
+
+```
+docs/05_Reports/session-report-2026-03-01-210241.md
+docs/05_Reports/session-report-2026-03-01-210822.md
+docs/05_Reports/session-report-2026-03-01-211526.md
+docs/05_Reports/session-report-2026-03-01-223747.md
+docs/05_Reports/session-report-2026-03-01-230159.md
+docs/05_Reports/session-report-2026-03-02-135835.md
+docs/05_Reports/session-report-2026-03-02-144854.md
+docs/05_Reports/session-report-2026-03-02-154638.md
+docs/05_Reports/session-report-2026-03-02-154857.md
+docs/adr/008-alert-to-infra-migration.md
+docs/adr/011-donation-outbox-infra-migration.md
+docs/adr/011-equipment-cache-infra-migration.md
+docs/adr/012-like-to-infra-migration-build-plan.md
+docs/adr/012-like-to-infra-migration.md
+module-app/src/main/java/maple/expectation/adapter/in/AlertPortAdapter.java
+module-app/src/main/java/maple/expectation/service/v2/LikeSyncService.java
+module-app/src/main/java/maple/expectation/service/v2/alert/DiscordAlertService.java
+module-app/src/main/java/maple/expectation/service/v2/alert/DiscordMessageFactory.java
+module-app/src/main/java/maple/expectation/service/v2/alert/dto/DiscordMessage.java
+module-app/src/main/java/maple/expectation/service/v2/auth/CharacterLikeService.java
+```
+
+## 검증 결과
+
+- ✅ PfastTest 통과
+- ✅ DoD 검증 통과
+- ✅ Stateless 설계 준수
+- ✅ 미완성 코드 없음
+
+---
+*자동 생성됨 by stop-validation.sh*
