+++ module-app/src/main/java/maple/expectation/service/v2/shutdown/EquipmentPersistenceTracker.java
@@ -9,6 +9,8 @@ import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import maple.expectation.core.port.out.PersistenceTrackerStrategy;
+import maple.expectation.core.port.out.PersistenceTrackerStrategy.StrategyType;
 import maple.expectation.infrastructure.executor.LogicExecutor;
 import maple.expectation.infrastructure.executor.TaskContext;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
