+++ module-app/src/main/java/maple/expectation/service/v2/shutdown/RedisEquipmentPersistenceTrackerAdapter.java
@@ -4,15 +4,16 @@ import io.micrometer.core.instrument.MeterRegistry;
 import java.time.Duration;
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
+import maple.expectation.core.port.out.PersistenceTrackerStrategy;
+import maple.expectation.core.port.out.PersistenceTrackerStrategy.StrategyType;
 import maple.expectation.infrastructure.executor.LogicExecutor;
 import maple.expectation.infrastructure.queue.persistence.RedisEquipmentPersistenceTracker;
 import org.redisson.api.RedissonClient;
 
 /**
- * Adapter wrapper for RedisEquipmentPersistenceTracker to implement v2 interface.
+ * Adapter wrapper for RedisEquipmentPersistenceTracker to implement core port interface.
  *
- * <p>This adapter bridges the core port interface to the deprecated v2 interface for backward
- * compatibility.
+ * <p>This adapter bridges the core port interface to the Redis implementation.
  *
  * <p><b>Phase 3 Technical Debt:</b> This adapter can be removed once all services are migrated to
  * use the core port interface directly.
