+++ module-app/src/main/java/maple/expectation/service/v5/worker/MongoDBSyncWorker.java
@@ -8,8 +8,8 @@ import jakarta.annotation.PreDestroy;
 import java.time.Duration;
 import java.util.Map;
 import lombok.extern.slf4j.Slf4j;
+import maple.expectation.core.event.ExpectationCalculationCompletedEvent;
 import maple.expectation.error.exception.InternalSystemException;
-import maple.expectation.event.ExpectationCalculationCompletedEvent;
 import maple.expectation.infrastructure.executor.CheckedLogicExecutor;
 import maple.expectation.infrastructure.executor.LogicExecutor;
 import maple.expectation.infrastructure.executor.TaskContext;
